StartPatch returns uint8_t, 0 on success, FR_TIMEOUT=15 on timeout

+++ firmware/patch.h
@@ -110,7 +110,7 @@ extern uint8_t i2crxbuf[8];
 extern uint8_t i2ctxbuf[8];
 
 void InitPatch0(void);
-int StartPatch(void);
+uint8_t StartPatch(void);
 uint8_t StopPatch(void);
 
 void start_dsp_thread(void);
